requirements.txt
amqp==5.1.1
arrow==1.3.0
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
binaryornot==0.4.4
black==23.9.1
celery==5.3.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.4.0
cron-descriptor==1.4.0
cssselect2==0.7.0
Django==4.2.5
django-appconf==1.0.5
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-compressor==4.4
django-filer==3.0.6
django-libs==2.0.3
django-livereload==1.7
django-livereload-server==0.4
django-polymorphic==3.1.0
django-timezone-field==6.0.1
easy-thumbnails==2.8.5
Faker==19.11.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
Pygments==2.16.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
rcssmin==1.1.1
reportlab==4.0.5
requests==2.31.0
rich==13.6.0
rjsmin==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.3
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
webencodings==0.5.1
